# api-documentation/api.yaml
swagger: "2.0"
info:
  description: "This is API documentation of CookIt."
  version: "1.0.0"
  title: "CookIt"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "daniorocket@gmail.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:5000"
basePath: "/api"
...
schemes:
- "https"
- "http"
paths:
...
  /v1/user:
    get:
        tags:
        - "auth"
        summary: "Get userinfo from JWT"
        description: "This require Authorization header with JWT. JWT must be valid. When data are valid password returned is always as empty string for security issues."
        operationId: "getuserinfo"
        consumes:
        - "application/json"
        produces:
        - "application/json"
        responses:
          "200":
            description: "successful operation"
            schema:
              type: "object"
              properties:
                data:
                  $ref: "#/definitions/User"
                status:
                  type: "string"
                  example: "success"
                code:
                  type: "integer"
                  example: 200
                error:
                  type: "string"
                  example: "none"
          "401":
            description: "Invalid token"
            schema:
              type: "object"
              properties:
                data:
                   default: null
                status:
                  type: "string"
                  example: "failed"
                code:
                  type: "integer"
                  example: 401
                error:
                  type: "string"
                  example: "Failed to authorize user"
        security:
        - token: []
...
  /v1/recipes:
      post:
        tags:
        - "recipe"
        summary: "Add recipe"
        description: "This method allow to add recipe"
        operationId: "addrecipe"
        consumes:
        - "application/json"
        produces:
        - "application/json"
        parameters:
        - in: "body"
          name: "body"
          description: ""
          required: true
          schema:
            $ref: "#/definitions/Recipe"
        responses:
          "200":
            description: "Successful operation"
          "400":
            description: "Bad request"
      get:
        tags:
        - "recipe"
        summary: "Get all recipes"
        description: "This method allow to get all recipes. Required keys page and limit for pagination"
        operationId: "getallrecipes"
        consumes:
        - "application/json"
        produces:
        - "application/json"
        parameters:
        - in: "query"
          name: "page"
          description: ""
          required: true
          type: "integer"
        - in: "query"
          name: "limit"
          description: ""
          required: true
          type: "integer"
        responses:
          "200":
            description: "Successful operation"
          "400":
            description: "Bad request"
          "401":
            description: "Unauthorized"
  /v1/recipes/tags:
    get:
      tags:
      - "recipe"
      summary: "Get all recipes by tags"
      description: "This method allow to get all recipe by tags. Required keys page and limit for pagination"
      operationId: "getallrecipesbytags"
      produces:
        - "application/json"
      parameters:
      - in: "query"
        name: "page"
        description: ""
        required: true
        type: "integer"
      - in: "query"
        name: "limit"
        description: ""
        required: true
        type: "integer"
      responses:
        "200":
          description: "Successful operation"
        "400":
          description: "Bad request"
        "401":
          description: "Unauthorized"
definitions:
...
  User:
    type: "object"
    properties:
        id:
          type: "string"
        username:
          type: "string"
        password:
          type: "string"
          example: ""
        avatarURL: 
          type: "string"
        email: 
          type: "string"
        description: 
          type: "string"
...
  Category:
      type: "object"
      properties:
        id:
          type: "string"
        labelPL:
          type: "string"
        labelEN:
          type: "string"
        file:
          $ref: "#/definitions/File"
  File:
      type: "object"
      properties:
        encodedURL:
          type: "string"
        extension:
          type: "string"
  Recipe:
    type: "object"
    properties:
      name:
        type: "string"
      listOfSteps:
        type: "array"
        items:
          type: "string"
          example: "a"
      listOfCategories:
        type: "array"
        items:
          $ref: "#/definitions/Category"
      tags:
        type: integer
      description:
        type: "string"
...
securityDefinitions:
  token:
    type: "basic"